updating code to make best use of the maze class
